refactor(tests): rename serializationErrors to seismicTxValidation
The tests verify seismic tx field-presence validation rather than
serialization format errors. Rename the file; function names unchanged.
