+++ clients/ts/packages/seismic-viem-tests/src/tests/seismicTxValidation.ts
